fix(timer): play button resets on finish + explicit CGFloat cast
- toggle() now resets the timer when isFinished=true so the play button
  is always actionable after a session completes
- Explicit CGFloat(timer.progress) cast for type clarity

+++ Sources/Cadence/Views/Timer/TimerPanelView.swift
@@ -33,7 +33,7 @@ struct TimerPanelView: View {
             // Timer clock — style driven by settings
             TimerClockView(
                 style: settings.timerStyle,
-                progress: timer.progress,
+                progress: CGFloat(timer.progress),
                 isFinished: timer.isFinished,
                 timeString: timer.timeString,
                 isRunning: timer.isRunning
